requirements-colab.txt
astunparse==1.6.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
expecttest==0.2.1
ffmpeg-python==0.2.0
filelock==3.15.4
fsspec==2023.6.0
future==1.0.0
hashin==1.0.1
huggingface-hub==0.24.0
hypothesis==6.108.2
idna==3.7
Jinja2==3.1.4
lintrunner==0.12.5
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
optree==0.12.1
packaging==24.1
pillow==10.4.0
pip-api==0.0.34
psutil==6.0.0
PyYAML==6.0.1
RealESRGAN @ git+https://github.com/sberbank-ai/Real-ESRGAN.git@362a0316878f41dbdfbb23657b450c3353de5acf
requests==2.31.0
setuptools==71.0.1
six==1.16.0
sortedcontainers==2.4.0
sympy==1.13.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
types-dataclasses==0.6.6
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.43.0
jedi>=0.16
